productizer/pyproject.toml
[tool.poetry]
name = "example-productizer-httpsig"
version = "0.1.0"
description = "Example Productizer supporing HTTP Message Signatures"
authors = ["IOXIO"]
packages = [{include = "app", from = "."}]
license = "BSD-3-Clause"

[tool.poetry.dependencies]
python = "^3.9"

flask = "^2.2.3"
pydantic = "^1.10.6"
python-dotenv = "^1.0.0"
http-message-signatures = "^0.4.4"
requests = "^2.28.2"
pyjwt = "^2.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
